Mac用にbashrcを修正。
Macの場合はviでMacVimが呼び出されるようにした。
ターミナル.appで.bashrcが読み込まれるように.profileを追加。
シンボリックリンクを作成するようにmake_link_mac.shに追加。全体的にファイルの存在をチェックするコードに修正している。

+++ make_link_mac.sh
@@ -1,7 +1,19 @@
 #!/bin/sh 
 
-ln -s ~/dotfiles/vimfiles ~/vimfiles
-ln -s ~/dotfiles/_vimrc ~/.vimrc
-ln -s ~/dotfiles/_gvimrc ~/.gvimrc
-ln -s ~/dotfiles/_bashrc ~/.bashrc
+if [ ! -d ~/vimfiles ]; then
+	ln -s ~/dotfiles/vimfiles ~/vimfiles
+fi
+
+if [ ! -f ~/.vimrc ]; then
+	ln -s ~/dotfiles/_vimrc ~/.vimrc
+fi
+if [ ! -f ~/.gvimrc ]; then
+	ln -s ~/dotfiles/_gvimrc ~/.gvimrc
+fi
+if [ ! -f ~/.bashrc ]; then
+	ln -s ~/dotfiles/_bashrc ~/.bashrc
+fi
+if [ ! -f ~/.profile ]; then
+	ln -s ~/dotfiles/_profile ~/.profile
+fi
